Move term scraper to separate file
First step in abstracting everything in term-scraper to be more easier
to compose. This will also eventually result in less duplication between
the term and the post scraper.
